+++ src/main/java/entities/Trabajos.java
@@ -6,7 +6,6 @@
 package entities;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 import javax.persistence.Basic;
@@ -43,7 +42,7 @@ public class Trabajos implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "ID", nullable = false, precision = 38, scale = 0)
-    private BigDecimal id;
+    private Integer id;
     @Size(max = 40)
     @Column(name = "EMPRESA", length = 40)
     private String empresa;
@@ -58,15 +57,15 @@ public class Trabajos implements Serializable {
     public Trabajos() {
     }
 
-    public Trabajos(BigDecimal id) {
+    public Trabajos(Integer id) {
         this.id = id;
     }
 
-    public BigDecimal getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(BigDecimal id) {
+    public void setId(Integer id) {
         this.id = id;
     }
